[auth-service]: Refactor Form implementation.
This is an attempt to refactor the way we render forms for the groups
page. Previously we only had to care about the EditGroupForm, now we
have added the NewGroupForm. This means that we have to refactor the
workflow from just clearing the text content of the HTML element to
keeping track of state and rendering the two different forms differently
as they have different use cases and implementations. This
implementation isn't perfect but is a step in the right direction to
making the groups page and auth service more modular.

Bug:1336134
Change-Id: I7a395754bb69bc26c8b092b87b3f14fd2125dda1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3825158
Commit-Queue: Camilo Jacomet <cjacomet@google.com>
Reviewed-by: Joey Scarr <jsca@google.com>