[auth-service]: Group form validation.
This handles validation for both types of Group Form classes (new and edit). It validates specific fields on input in that field, and also performs validation on all fields when submit event is triggered. Currently does not actually submit the form, just logs that the form is valid to submit.

Change-Id: Ie67a0dcf8a512fdee5e40f5428e61c653468379a
Bug: 1336134
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3834727
Commit-Queue: Camilo Jacomet <cjacomet@google.com>
Reviewed-by: Joey Scarr <jsca@google.com>