[milo] De-emphasize durations on build page
The duration badges had the same styling as clickable buttons and this
either made them look clickable or made the actual buttons look like
badges. It also causes them to stand out and for most steps, which take
on the order of milliseconds, it's not important information.

This change gives the duration badges a light color and lighter font,
giving them a de-emphasized look. In addition, the lightness varies
based on the duration units, so milliseconds are lighter than seconds
which are lighter than minutes, etc. This allows for quickly scanning
the page for slow steps, which is common when debugging a slow build.

http://screen/Bhmm8ZQHWBNJTRX (before)
http://screen/6KL8RkHNxDfFEct (after)
http://screen/4hBumeeAgJniSgj (after, test results)

Change-Id: I8111f290ecfeb02a3898fb3b6f4a68f6fadda0c0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3846014
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Erik Staab <estaab@chromium.org>