[milo] change how popup components' styles are applied
Instead of using a single global cache, provide an option to render the
popup to the parent component in the same shadow DOM. So the style
applied to the rest of the component can naturally be applied to the
popup.
This is beneficial when the component contains a mix of pop ups and
regular components. With the previous strategy the styles will not be
applied properly.

R=mdraz@google.com, nqmtuan@google.com

Bug: 1357464
Change-Id: I68dc41b68b2aa49ce267a28d39543f00d6584a92
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3873266
Reviewed-by: Mostafa Draz <mdraz@google.com>
Auto-Submit: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Weiwei Lin <weiweilin@google.com>