[ResultDB] Optimize a hotspot in testvariants.ToTestResultProto.
This adds StringPairFromStringUnvalidated, a faster alternative to
StringPairFromString.  It can be safely used on data that was already validated
(such as data in our Spanner tables, which is validated before being written).
StringPairFromStringUnvalidated is now used in testvariants.ToTestResultProto,
which shows up in profiling as a significant hotspot.

Bug: 1359201
Change-Id: Ifcbd447f255f3ce94d5865d3589f589ac5637f7c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3871231
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Andrew Bennetts <abennetts@google.com>