[milo][history] fix filter text from querystring gets discarded
Previously, we check whether the test history page state is initialized
to decide whether to populate the filter with the filter setting from
the querystring. Now the test history page state is always
initialized. This causes the filter text from the querystring to be
always discarded.

This CL fixes the issue by stop relying on the state of this.pageState
to decide whether this.initialFilterText should be respected.

Unit tests will be added in a later CL so we can get the fix out
quickly.

R=mdraz@google.com, meiring@google.com

Change-Id: Ief7857313a8a3deaa15951d6fd83df0b9dc01c2e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3886276
Auto-Submit: Weiwei Lin <weiweilin@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>