[server] add flags for tsmon flush interval and timeouts
CV added a number of aggregated metrics, which walks through all the
project configs and reports the status of all the existing tryjobs
and runs.

As a result, the number of cells to be reported by a flush were
increased by ~10x, and CV's flush started timing out often.

There are a couple of different solutions for this problem, but
this CL is to make the flush timeouts configurable so that
CV can give out a longer timeout value as a short-term solution.

Change-Id: I772176633ca36b828f2a649487b41609662e10d2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3888520
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Scott Lee <ddoman@chromium.org>