[LUCI Analysis] Comment when merging rules for duplicate bugs.
LUCI Analysis's updating of rules when bugs get merged is currently
invisible to the user. Often the fact the bug has a rule is attached
at all after such a merge is not known.

Send a comment when merging rules to surface that LUCI Analysis
has done this.

BUG=None
TEST=Integration tests

Change-Id: If595680b6edf2516a6618eeccb699355afd80103
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3891353
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>