cv: override the tsmon state in the requst to report aggregation metrics
The code before has a bug that it is not using the new state created
but rather use the existing state in grpc server. So the flush still
follows the existing asynchronous workflow. This change fixes the bug
by overriding the state in the request context.

R=ddoman

Bug: 1362701
Change-Id: If4c336515d009f07a621871c278ba588ebbb50b7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3897417
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Auto-Submit: Yiwei Zhang <yiwzhang@google.com>