cv: Change BB listener to also parse ancestor_ids if available
Using bbv1pb.LegacyApiCommonBuildMessage was limiting because in order
to access the field ancestor_ids, the structs in
luci/common/api/buildbucket/buildbucket/v1/buildbucket-gen.go
would have to be regenerated based on the latest Buildbucket legacy
API changes, which wasn't working for me.

Instead of having to use that, we can define a private struct in this
package containing only the fields that we will use in CV. At the
moment, we only use hostname and build ID, but for crbug.com/1346460
in the next CL we'll want to use ancestor IDs.

Bug: 1346460
Change-Id: Icccb5ebfd8340d6d2139ca8bbe12a83ac4b0d2db
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3915044
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>