[auth] Annotate token refresh errors.
This is especially important when `ctx` associated with
the authenticator expires. A generic "context canceled" error
appearing in e.g. pRPC error message is very confusing.

Also switch to using errors.Annotate or errors.Reason instead of
fmt.Errorf.

Note that annotating all errors indiscriminately will break the
public API of this method, so don't annotate ErrLoginRequired.

ErrBadCredentials was another error that was documented as being
returned as is, but it is not checked by anything, so just remove
it from the public API.

R=iannucci@chromium.org, ddoman@chromium.org

Change-Id: I73ee193bd96c7ef6471d4bdc4c4aa07310dc9f6c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3937913
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>