cv: do not delay purging if there is no cq vote.
The purging can be delayed if the config allows combining CLs and has a
stabilization period. Such stabilization is added to the most recent cq
vote timestamp, but if there is no such vote, as in the case of CLs with
only a new patchset run, then we shouldn't try to find such a timestamp.

Also, adding test for this scenario.

R=yiwzhang

Change-Id: I4504a0bfd84a20a3fae4571a8f84ea3c52335131
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3938711
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>