[logdog] Switch `logs` to use custom Docker image.
GAE Flex builder is stuck at go1.15 runtime and it doesn't look like
it will ever be updated. The recommended solution is to bring your
own Docker image and use `runtime: custom`.

If we do that, we still need to give `module-logs.yaml` to the
deployer, since this YAML still defines various bits of GAE
configuration (routes, resources, scaling, etc.). But we don't
need any other file from the Logdog source code tree in the
GAE tarball, since the deployer won't actually be building this
code for us (instead we do that ourselves when building the
docker image).

We can reuse the big tarball with all the source code, and just
read one YAML from it, but this tarball will changing relatively
frequently (when any of the code changes), which will result in
redundants redeploys (`logs` module will be updated anytime
the tarball OR the docker image changes, where only the docker
image changes actually result in new code running).

So make a new "tarball" with just one `module-logs.yaml`. It
will be changing approximately never, which means we'll redeploy
`logs` module approximately only when the docker image changes,
which is what we want.

Also add some more info to README's.

Also with this change, deploying `logs` module from the local
checkout is not really supported (it is possible to do, but there
are no ready scripts that can do that and I don't think documenting
the manual process is worth the effort, since GAE Flex is
the evolutionary dead end).

R=iannucci@google.com

Change-Id: Ide94cb7de0fd5abd706ae3dbc1ac2849f8914f8a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3943229
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>