[cv] implement project_finder
projectFinder{} is written for the following intents.

1) cache the compiled regexes
pf.lookup() will be called for every single pubsub message.
There is no reason to compile them again and again.
2) decouple the logic of looking up the projects from gerritProc{},
so that
- the state of gerritProc{} stays the same, even if there is a change
in gobmap and svccfg.EnabledProjectRegexps.
- Hence, subscribers would never have to restart, unless there was
a change in the ReceiveSetting in listener-settings.cfg, which would
be quite rare.
3) makes it easier to add an optimization projectFinder{}, as the code
is isolated from gerritProc{}.

Bug: 1358208
Change-Id: Ic6765c47607a44800d8a7b561c8d74c4e5a47ebb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3932464
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>