[go] Stop using some ioutil functions deprecated in go1.16.
Logdog `logs` module was the last place where we used go1.15 and it
has been switched to a newer go version. Now go1.16 is the oldest
Go version we support (it is used on GAE Standard).

This removes usage of deprecated ioutil aliases:
  ioutil.Discard => io.Discard
  ioutil.NopCloser => io.NopCloser
  ioutil.ReadAll => io.ReadAll
  ioutil.ReadFile => os.ReadFile
  ioutil.WriteFile => os.WriteFile

There's also ioutil.ReadDir, but it is not a simple alias. Its usage
will be removed in a separate CL.

R=iannucci@chromium.org

Change-Id: Ibb32be30b9be8a269360b6ece2586513aafa1cba
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3946103
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>