[buildbucket] Change BuildFieldVisibility extension number.
"50001" is conflicting with an existing extension when importing
this proto into google3, see cl/481029678.

Pick the field number randomly and check it is not mentioned in
any *.proto in cs/. The official suggestion is to use Piper CL
number, but we are not in Piper. Random unused number should be
as good.

R=chanli@chromium.org

Change-Id: I55506037aed3919fb482e20981d799ac10590341
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3957215
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>