[LUCI Analysis] Add failure dispersion criteria to avoid filing for CL-specific failures.
A bug can now only be filed for a cluster if one of the two
following criteria are met:
- A failure in the cluster was observed in a postsubmit build.
  (Build testing a version of source in the tree and not a patch.)
- Failures in the cluster were observed over at least 3 distinct
  user CLs (i.e. not automation-generated uprev CLs).

Approximately 13 bugs which were filed by LUCI Analysis over the last week would not have been filed if it were for these additional criteria.

We will need to consider how we surface this dispersion criteria
on the UI to make sure the system's bug filing criteria remains
understandable.

BUG=b:249435968,b:250541091
TEST=INTEGRATION_TESTS=1 go test go.chromium.org/luci/analysis/...

Change-Id: I0f4650a1ce4902ba28d02a86709583eedecc5add
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3954805
Reviewed-by: Mostafa Draz <mdraz@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>