[cipd] Remove strong dependency on a gRPC server from the client.
There are issues when importing a gRPC server registration into
google3 code. Some extremely obnoxious system discovers it via
static code analysis and insists on exposing this server in some
Goa configs, which look cryptic to me and not something I want to
maintain.

It happens because CIPD client code in google3 is linked into
some server (via some code-we-dont-own) which is a part of boq/goa
universe. The goa compliance checker stupidly assumes that any
gRPC servers linked into a binary (even transitively, even in
third_party, even when not instantiate) are part of the server
API and must be separately described in some cryptic (to me)
manifest files.

CIPD plugins currently only matter when installing packages.
google3 CIPD client explicitly doesn't support installation of
packages, so just exclude the plugin system (and the gRPC server)
from google3 export.

This was done previously in [1] but was accidentally reverted in [2].

[1]: https://chromium.googlesource.com/infra/luci/luci-go/+/82a0ac409
[2]: https://chromium.googlesource.com/infra/luci/luci-go/+/82d549a81

R=iannucci@chromium.org

Change-Id: Ia20162d5c836f048d7f637f2d1d432b50818dcae
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3961728
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>