[LUCI Analysis] Add QueryExoneratedTestVariants RPC.
This new RPC will support a test exonerations section on the UI.
Use of a separate RPC ensures that this section continues to function
correctly even if there are many recent failures in the cluster
which are from test variants that are not exonerated.

The UI will chain this RPC with the existing QueryFailureRate RPC.
Combining the RPCs was considered but not done due to API design
guidance suggesting that combination methods can lead to explosions
of methods/unmaintainable API surface and it would only avoid one
server round-trip.

Ideally this method would be a more general purpose method for
obtaining test variants (e.g. a standard AIP-132 List method with
filter and order_by support) but for implementation simplicity
this was not done. We can migrate to such a solution in future.

BUG=b:253695379
TEST=Integration tests

Change-Id: Ia91882c9ef74bb9a46d8c35b9e8e62529fe4133d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3977051
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Matthew Warton <mwarton@google.com>