[ResultDB] Populate the fetchURL with GcsURI for artifact stored in GCS.
Attempt to populate the fetchURL with GcsURI if one is available.
ResultDB currently provides 3 endpoints to fetch artifact(s):
- Get Artifact
- List Artifacts
- Query Artifacts

Modified Query and Read so that GcsURI is read and fetchURL could be
generated from it.

This approach does not require additional Spanner lookup which caused
errors due to timeout in the previous attempt.

BUG=b:229888940
TEST=Manual Unit Test

Change-Id: I0c25300e00f00a6ca732c41279749baa265c2b1c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3969709
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Jason Kusuma <jkusuma@google.com>