[milo] fix files with special characters are not cached properly
The previous solutions APPEARs to work (but doesn't) because all chunk
files with special characters are included in the HTML file.
As soon as one of them is not included, the page breaks because webpack
itself doesn't encode the filename when dealing with async imports.

The fundamental issues is that html-webpack-plugin encodes chunk
filenames when the rest of the tool stacks (including workbox and
webpack itself) don't expect the filenames to be encoded.

R=mdraz@google.com, nqmtuan@google.com

Change-Id: If797d338a77e7225cc1d2e3c9b23fa09b465e77a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3995656
Auto-Submit: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Weiwei Lin <weiweilin@google.com>
Reviewed-by: Mostafa Draz <mdraz@google.com>