[led] do not put the recipe bundle in "kitchen-checkout" dir if bbagent will do the package download in the led build
In today's led jobs, the swarming task is responsible to download cas inputs and it downloads the inputs to {TASK_ROOT}. So in `led recipe-bundle`, the bundle
is put in the "kitchen-checkout" dir (since "kitchen-checkout" is the hardcoded payload path).

In the builds that bbagent is responsible to download cas inputs, bbagent will download the inputs to the desinated path as `build.infra.bbagent.input.data` indicates. Which means it will put the bundle in {TASK_ROOT}/kitchen-checkout already. So in this case the bundle should not be put in "kitchen-checkout" dir when is upload to CAS.

Bug:1114804
Change-Id: Ic55e42a47f37a8b9a24cfb7ab4607103612581c9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3991687
Reviewed-by: Yuanjun Huang <yuanjunh@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Chan Li <chanli@chromium.org>