[secrets] Always read previous version of secrets from GSM.
This removes `withPrevVersion` in internal readAndTrackSecret(...)
API in favor of always doing it to simplify code and avoid
surprises with StoredSecret API which claims to return
previous versions but in reality doesn't.

It was essentially "an optimization" for Tink secrets because
it was believed it is possible to do a Tink key rotation purely
by updating the keyset in the "latest" version of the secret.
In the new plan we'll assemble the final keyset from a bunch of
individual secret versions, and thus we will need to load them.
This will be overall less complexity in the end, I promise.

Practically, this CL will cause ~2x QPS to AccessSecretVersion
API, since we will now try to load previous versions of all
stored secrets (Tink keys, but also OAuth client secrets).
This should be OK though, since even in busy services we
currently use only 0.01% of available quota for
AccessSecretVersion (because we cache secrets heavily in
the process memory).

R=iannucci@chromium.org
BUG=1380939

Change-Id: I8ee36d7f2192602cd8ebb3a5fb5bccda30e229a1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/3998987
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>