[auth] Forbid interactive flow if SWARMING_HEADLESS in set.
This is a precaution against running it on bots, since this is now
more dangerous (one can take the login link from the bot log and
try to finish the flow ignoring all the warnings, which will result
in some end-user credential on the bot machine).

Also bump the version, since I plan to release this to depot_tools.

R=iannucci@chromium.org

Change-Id: I33a1cff17ca1266e6c6179e50f2078dd49952bf4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4004540
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>