[milo] fix rerendering of artifact tags in summary html
Because lit-html is unable to compare two TrustedHTML objects, it
rerenders the summary_html everytime the parent component rerenders.
This causes the artifact tags to be refreshed and reloaded.

The solution is to pass string instead of TrustedHTML to unsafeHTML,
and rely on the default policy to sanitize the input and convert it
to TrustedHTML.

Because milo needs to support artifact tags, milo needs to define its
own default policy rather than relying on the shared trusted-types
-policy package.

R=mdraz@google.com, nqmtuan@google.com

Bug: 1381900
Change-Id: I951a27c6baeb095424dc4f06bd2e46bba0aac61a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4005616
Reviewed-by: Mostafa Draz <mdraz@google.com>
Commit-Queue: Mostafa Draz <mdraz@google.com>
Auto-Submit: Weiwei Lin <weiweilin@google.com>