[milo] make an expandable entry in react
So it can be used in react components.
The original lit-element component is not converted to a wrapper around
the new react component because
1. It will be difficult to make it react to events coming from shadow-
dom. This is required to make the entry expand when user clicks on the
injected header component.
2. It could make the rendering performance slower. The could be >10000
expandable entries rendered on the page.

R=mdraz@google.com, nqmtuan@google.com

Bug: 1380615
Change-Id: If6063838f6f38130f30a96f43f6213e1b8de6ca5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4005605
Commit-Queue: Mostafa Draz <mdraz@google.com>
Auto-Submit: Weiwei Lin <weiweilin@google.com>
Reviewed-by: Mostafa Draz <mdraz@google.com>