[Analysis] Move to project-scoped auth where possible.
Authenticate to buildbucket and ResultDB using project-scoped
credentials where possible. This is notionally more secure against
confused-deputy type security bugs and will remove the need for
individual projects to authorise LUCI Analysis to access their data.

One exception needs to be made for checking if the ancestor build
contains the build being considered for ingestion. This access
cannot be made using the project-scoped credentials as the
ancestor build may be in a different project to the ingested build.

BUG=b:258980021
TEST=Integration tests

Change-Id: Iff3abc02dc748595da0346177e2de925388671d3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4029559
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>