[Bisection] Introduce runstatus + correctly set analysis status
With the current status, for some cases, we won't know if an analysis is running. For example, if an analysis status is SUSPECT_FOUND, has it already stopped, or it still waiting for some verification runs?

I've introduced the RunStatus in this CL to focus on run status, rather than result status.

Also as we are running both heuristic and nthsection, the analysis status must be updated to reflect the status for both sub-analysis.


Bug:1295566
Change-Id: I2c4fcf552960bf3da722cd726f1b35fc48bf69d4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4040801
Reviewed-by: Anne Redulla <aredulla@google.com>
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>