[cv] update configs/prjcfg/loader_test.go to remove dep on refresh.go
A follow-up CL will move refresh.go to refresher/ but it causes
an import cycle, as loader_test.go uses functions in refresh.go.

This CL replaces the use of UpdateProjectConfig() and
DisableProjectConfig() with own mock functions.

1. loader.go doesn't really use any of refresh.go functions.
They simply fetch and return the stored entities.
loader_test.go uses refresh functions for convenience.

2. It's not necessary that the mocked refresh functions match
the exact same behaviour of the original refresh functions.
Due to the same reason described in (1). loader_test.go
is not to test if the refresher functions update entities
as expected.

Bug: 1358208
Change-Id: Idafb9544824638ca1a0cdf2fdfd87d6059a91b3f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4048785
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Scott Lee <ddoman@chromium.org>