[lucicfg][doc] Minor updates
Add the word "quota" to the description of the user_limits argument to
luci.cq_group(). The word "quota" currently only shows in the arguments
to cq.retry_config() but is likely to be used to search for the
functionality of user_limits.

Make minor changes to a references to cq.user_limit where linking
conflicted with fixed-width formatting.

Also it appears README.md wasn't regenerated after a recent change so
there's a minor update on location_filters.

Change-Id: I486db0102b460d5ae088da22900b7ce41daa393a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4086901
Auto-Submit: Rob Mohr <mohrr@google.com>
Commit-Queue: Scott Lee <ddoman@chromium.org>
Reviewed-by: Scott Lee <ddoman@chromium.org>