[bbagent] Set environment variable when python2 is omitted from build.
Environment variable is:

    LUCI_OMIT_PYTHON2=true

This will be set when the 'luci.buildbucket.omit_python2' is set for
the build.

This environment variable will be used by e.g. depot_tools to determine
when it absolutely should not use python2, even if it would otherwise
find one in PATH (e.g. from leaked system installs of python).

This also makes the python2 omission experiment imply usage of python3 for the recipes.

R=chanli, randymaldonado, sokcevic, tikuta, yuanjunh

Bug: 942720, 1359438
Change-Id: I0fb4f731d41801e9611fd4df199a359f85aaa902
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4087937
Reviewed-by: Josip Sokcevic <sokcevic@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Randy Maldonado <randymaldonado@google.com>