resultdb: Attempt to optimiza reachability graph traversal
The reachability graph traversal has been profiled to take more than 40
seconds in some cases.  Simplify the code and get Spanner to do the work
in one query per graph level rather than one query per graph node.

Only update QueryTestVariants and BatchGetTestResults in this CL, other
uses of graph traversal will be updated subsequently.

The Redis cache is not used for this CL.  It is expected that traversal
is fast enough that it should not matter.  It can be added back in in a
subsequent CL if necessary.

When this is rolled out to Canary, the performance of the two affected
RPCs has to be examined before doing a full prod roll out.

Change-Id: Ibb4867c2cb215fea4dabc2e2776c368731741b47
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4105420
Commit-Queue: Matthew Warton <mwarton@google.com>
Auto-Submit: Matthew Warton <mwarton@google.com>
Reviewed-by: Patrick Meiring <meiring@google.com>