[Analysis] Put in place framework for metrics (part 1).
Begin putting in a framework for metrics that will make it
simpler to add more metrics. This refactoring aims to
remove the hardcoding of per-metric fields and logic in protocol
buffers/RPC handlers, SQL statements/analysis and the UI. Instead,
metrics will have by a named metrics.Definition structure and
all logic will be generic to the metric, identifying it only by
its ID (name).

This initial change refactors the part of the backend and changes
the format of the cluster summaries table but does not make changes
to any protocol buffers. These will follow in later changes.

BUG=b:260631527
TEST=Integration tests

Change-Id: Ib6a033fc3d3c4ba884a6f33fb92256b32d9f661a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4060956
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>