[luciexe] Add support for LUCIEXE_FAKEBUILD
This change adds support for LUCIEXE_FAKEBUILD to luciexe binaries. It
accepts a path to a jsonpb Build proto and sets up the luciexe binary
according to the LUCI User Code Contract.

It also adds the --working-dir flag to luciexes to specify which
directory to run out of; this is useful for the fake build case because
it lets us clean up the temporary directory it runs in by default.

The purpose of this feature is to make it easier to run luciexe binaries
locally.

Change-Id: I4bd12093c0087ceb1ad19a44d20cb9214dfb4e36
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4098465
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>