resultdb: disallow credentials on CORS requests
The indicated fallback to OAuth2 authentication for artifact content
requests was never implemented.  This proves it is never used.

Given this is not needed, we can disallow
credential (such as cookies) transmission on CORS requests.  This means
we do not need the Origin header from the request and can just use
Access-Control-Allow-Origin: "*" (which is effectively what the old code
was doing anyway).

Why do this?  A new client is loading images from ResultDB using <img
src="..."> tags from a different origin.  However, these requests do not
send an Origin header, so no CORS headers are added to the response.
When the page subsequently tries to access the image data (doing pixel
diffing) the browser refuses access due to CORS tainting.

Bug: 1408855
Change-Id: If6dde6ac46822edf682ca7dce6e5b8c982d498fd
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4189288
Reviewed-by: Mostafa Draz <mdraz@google.com>
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Matthew Warton <mwarton@google.com>