[Bisection] Fix analysis status not being set correctly
For example, in https://luci-bisection.appspot.com/analysis/b/8790929965715736625, the analysis suspect was not set correctly (it should be FOUND, not SUSPECT_FOUND, because the nthsection suspect has been verified).

It was because when we update analysis status, LUCI Bisection didn't check if there is any suspect scheduled to be verified (in the task queue), and mistakenly assumed that the analysis has finished because there was no rerun pending/running.


Bug:1295566
Change-Id: Ie4d337c3d49164dc53cdc90d5204fff2a8f3f452
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4199858
Commit-Queue: Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Anne Redulla <aredulla@google.com>