[cas] Follow symlinks to files outside working dir
`cas archive` now works for symlinks that point outside of the working
directory; previously it would fail with a message like "path
/a/b/otherdir/foo is not under /a/b/workdir".

This makes it much easier to use, as we no longer need recipe logic to
copy files into a temporary directory before uploading to CAS, and can
instead upload in-place. It also makes the CAS client consistent with
the old Isolated client, which implemented this behavior.

Fixed: 1216363
Change-Id: Ia16dc4efe1ef9ce3b7c5224c0c991bf516d7cd37
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4027826
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Oliver Newman <olivernewman@google.com>