[common/proto/msgpackpb] Add a library for encoding proto as msgpack.
This will be used by the server/quota implementation to communicate
bidirectionally with Redis lua scripts in a type safe way.

Redis's lua interpreter has a fast implementation of msgpack, but no
implementation of protobuf. Since we want to optimize the size of the
data at rest, this encoding scheme is relatively space efficient.
Testing with a few of the pertinent messages from server/quota, this encoding seems to be roughly 7x smaller than the equivalent JSON
encoding.

I also didn't want to entirely give up on protobuf (as an IDL); earlier
versions of this library hand-rolled encoding methods, but not only was
this painful to maintain and change as the library implementation
improved, it was difficult to imagine how schema migrations would work
for data at rest.

R=ddoman, vadimsh

Bug: 1359069
Change-Id: Ibdb517777f8e8f58c6c1b68de748e6937cc40d86
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4100804
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Scott Lee <ddoman@chromium.org>