[common-data] Add writeByte and encode functions.
writeByte(dst, offset, bitOffset, val) writes the byte val to
the dst[offset] (and possibly dst[offset+1]) inside the dst buffer
according to the rules of base128 encoder. The bitOffset tells you
the distance from the start of dst[offset] to write to.

For example, 0b_1111_1111 written might get written as:

0b.111_111 0b.1**_***

In the notation above, a "." denotes a zero that is required to
be zero in the output, a "*" represents zeroes in the next byte
that we don't touch when encoding the current byte.

encode(dst, src) works by walking the source buffer one byte at
a time and repeatedly invoking writeByte to write the data from
src in the low seven bits of each byte in dst.

BUG=b:266749911
TEST=make test

Change-Id: Ic8f55ef3fbf7456fb19613db288932c83c0df976
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4199139
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>