[common-errors] Add a Append(...) function to multierror.
Creating a new perhaps-MultiError that is flat (i.e. nil or
an ordinary error) whenever possible is kind of tricky.

Provide a function that supports this idiom. The intended way
to use errors.Append is:

> err := DoSomething()
> err = errors.Append(err, DoSomethingElse())
> err = errors.Append(err, DoAThirdThing())
>
> if err != nil {
>   // log an error or something, I don't know
> }
> // proceed as normal

The `err != nil` check above is actually correct since we will
flatten an error to nil if possible.

BUG=b:266743374
TEST=unit tests

Change-Id: If60040826d6340bc86bc1437b481f0a70dde4a48
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4194040
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>