CV: Do not break on expected config hash mismatch.
If a config is changing in the same transaction where runs using this
new config are being created, expect the mismatch and do not fail the
task.

Use the age of the offloaded project config to determine if the mismatch
is due to the batching of the run creation and config update and not
because the config changed independently of the ManageProject task.

R=yiwzhang

Bug: 1400498
Change-Id: Ib25270922660b5a8279c8dcb2fcf86d07585e56b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4103913
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>