[common/errors] Correctly render regular Go errors implementing Unwrap.
Previously Unwrap was mostly implemented by LUCI errors... however the
stdlib adopted an identical interface (yay!) which ended up triggering
a bug in the error rendering logic. Specifically, the error rendering
logic buffers all wrappers until it encounters a LUCI stackframe
(stackContexter) in the error. Upon encountering a stackContexter, it
will then print all of the encountered wrappers in the context of that
stack frame.

However, this rendering logic failed to account for cases where the
error is either entirely an external error, or is an annotation of an
external error and would end up fully unwrapping an error, but then
discarding all of that wrapper information.

Now the rendering logic will keep a pointer to the last stackContexter
it actually encountered. Once it finishes unwrapping the error all the
way to the bottom, it will render the original error from the point in
the error tree where it last saw a stackContexter. This also means that
for pure Go errors (with no stack context at all), they will now be
rendered without the `original error:` preamble which is part of the
stack trace display.

R=vadimsh

Change-Id: Ifd005c52e8407c9ff90a9fe6b5dd320e56d7d831
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4229612
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>