[grpcmon] Add grpcmon StreamServerInterceptor.
It is copy-pasta because in the end state it will be sufficiently
different from the unary interceptor (and for that reason we should
not expose it as a UnifiedServerInterceptor).

The thing that will make it different (counting number of received
and sent packets) is not implemented yet though.

R=iannucci@chromium.org

Change-Id: Ie9220cca9ab6265f3f9494e4604f440471d8c859
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4235232
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>