[server] Refactor code that initializes per-request context.
Make it not depend on *http.Request that much. This allows to
reuse it for gRPC handler.

One small caveat is that Stack driver APIs (like LogEntry proto
and span metadata) are HTTP-only, so we'll need to do some URL
"reconstruction" for gRPC handlers to log them as if they are
just HTTP/2 ones.

R=iannucci@chromium.org

Change-Id: Iad2e1e5cc414ce3898e695f014cdbb2daedbc9de
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4237768
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>