[auth] Fix CLI login when scopes don't include `userinfo.email`.
This happens, for example, when using `luci-auth login -scopes-iam`.

It regressed after switching to the login sessions protocol because
some code there always expects an ID token, which is missing if
requested scopes do not include `userinfo.email` or `openid`.
The error message is `the authorization provider didn't return
ID token`.

The ID token checks are not critical and I opted to just remove them.
An alternative is to always append `openid` scope to the list of
scopes, but this change will be more visible.

R=iannucci@chromium.org, cjacomet@google.com

Change-Id: I5517df564ce34e72905daeb6bcccf9bf93658a10
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/4246945
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Camilo Jacomet <cjacomet@google.com>